Add revised version of additional files
The original manuscript, model codes and other supplementary materials were submitted to Environmental Health Perspectives on June 14, 2020. We received the decision letter and reviewers' comments on September 1, 2020. We revised the model code slightly based on reviewers' comments. This is the revised version of the model codes in the "Additional files" zip file as of October 27, 2020.